test(gcode-import): split page scenarios
